Refactor JS to enable footer to be at bottom of page
This is such a fucking hack job ... I am sorry future Andy. Really
sorry.

The basic issue is this:
- there is some custom JS that is used to resize the `#targetdiv`
element on the page on route change, which is supposed to set the height
of the scrolling background image element so the image isn't too tall.
- This is .... fine ... for the front page, but this #targetdiv element
is also where all the other pages render inside of. This height setting
functionality is only needed on the home page, but all pages will have
their target div resized on load
- What this results in is that the footer will not be at the bottom of
the page, but in the middle of the rest of the text content. Because
we're setting inline height styles on the target div, the actual flow of
the document is disrupted

My shitty, Friday-at-6:45pm solution to this is to look at the URL on
page load and only resize the target div for the homepage. This should
be changed in the future... but it will not be changed today.