Refactor header images
We've added this anniversary logo, which necessitated a new layout for
the header. I refactored from bg images to img tags because I now need
to lay these things out in a flexbox to align them. Moved that wacky
anchor tag outside the header logo to make the whole thing clickable and
to make the flexbox layout easier to implement. Removed a bunch of
hard-coded height styles to make it easier for the whole thing to flow
naturally.

I've learned a lot about CSS since I did this thing originally 😅